Add a test class for non streaming order by unit tests
