+++ Microsoft.Azure.Cosmos/tests/Microsoft.Azure.Cosmos.Tests/Query/Pipeline/OrderByCrossPartitionQueryPipelineStageTests.cs
@@ -6,6 +6,7 @@ namespace Microsoft.Azure.Cosmos.Tests.Query.Pipeline
 {
     using System;
     using System.Collections.Generic;
+    using System.Diagnostics;
     using System.Linq;
     using System.Threading;
     using System.Threading.Tasks;
@@ -21,6 +22,7 @@ namespace Microsoft.Azure.Cosmos.Tests.Query.Pipeline
     using Microsoft.Azure.Cosmos.Query.Core.Pipeline.CrossPartition.Parallel;
     using Microsoft.Azure.Cosmos.Query.Core.Pipeline.Pagination;
     using Microsoft.Azure.Cosmos.ReadFeed.Pagination;
+    using Microsoft.Azure.Cosmos.Serialization.HybridRow;
     using Microsoft.Azure.Cosmos.Tests.Pagination;
     using Microsoft.Azure.Cosmos.Tracing;
     using Microsoft.Azure.Documents;
@@ -631,229 +633,6 @@ namespace Microsoft.Azure.Cosmos.Tests.Query.Pipeline
             }
 
             Assert.AreEqual(numItems, documents.Count);
-        }
-
-        [TestMethod]
-        public async Task TestNonStreamingOrderByPipeline()
-        {
-            IDocumentContainer innerDocumentContainer = await CreateDocumentContainerAsync(10);
-            IDocumentContainer documentContainer = new NonStreamingDocumentContainer(innerDocumentContainer);
-
-            int pageSize = 10;
-
-            TryCatch<IQueryPipelineStage> monadicQueryPipelineStage = OrderByCrossPartitionQueryPipelineStage.MonadicCreate(
-                    documentContainer: documentContainer,
-                    sqlQuerySpec: new SqlQuerySpec(@"
-                        SELECT c._rid AS _rid, [{""item"": c.pk}] AS orderByItems, c AS payload
-                        FROM c
-                        WHERE {documentdb-formattableorderbyquery-filter}
-                        ORDER BY c.pk"),
-                    targetRanges: await documentContainer.GetFeedRangesAsync(
-                        trace: NoOpTrace.Singleton,
-                        cancellationToken: default),
-                    partitionKey: null,
-                    orderByColumns: new List<OrderByColumn>()
-                    {
-                        new OrderByColumn("c.pk", SortOrder.Ascending)
-                    },
-                    queryPaginationOptions: new QueryPaginationOptions(pageSizeHint: pageSize),
-                    maxConcurrency: 10,
-                    continuationToken: null);
-
-            Assert.IsTrue(monadicQueryPipelineStage.Succeeded);
-
-            IQueryPipelineStage stage = monadicQueryPipelineStage.Result;
-            List<CosmosElement> documents = new List<CosmosElement>();
-            while (await stage.MoveNextAsync(NoOpTrace.Singleton, default))
-            {
-                Assert.IsTrue(stage.Current.Succeeded);
-                documents.AddRange(stage.Current.Result.Documents);
-            }
-
-            Assert.AreEqual(10, documents.Count);
-        }
-
-        private sealed class NonStreamingDocumentContainer : IDocumentContainer
-        {
-            private readonly IDocumentContainer inner;
-
-            public NonStreamingDocumentContainer(IDocumentContainer inner)
-            {
-                this.inner = inner ?? throw new ArgumentNullException(nameof(inner));
-            }
-
-            public Task<ChangeFeedPage> ChangeFeedAsync(
-                FeedRangeState<ChangeFeedState> feedRangeState,
-                ChangeFeedPaginationOptions changeFeedPaginationOptions,
-                ITrace trace,
-                CancellationToken cancellationToken)
-            {
-                return this.inner.ChangeFeedAsync(feedRangeState, changeFeedPaginationOptions, trace, cancellationToken);
-            }
-
-            public Task<Record> CreateItemAsync(CosmosObject payload, CancellationToken cancellationToken)
-            {
-                return this.inner.CreateItemAsync(payload, cancellationToken);
-            }
-
-            public Task<List<FeedRangeEpk>> GetChildRangeAsync(
-                FeedRangeInternal feedRange,
-                ITrace trace,
-                CancellationToken cancellationToken)
-            {
-                return this.inner.GetChildRangeAsync(feedRange, trace, cancellationToken);
-            }
-
-            public Task<List<FeedRangeEpk>> GetFeedRangesAsync(ITrace trace, CancellationToken cancellationToken)
-            {
-                return this.inner.GetFeedRangesAsync(trace, cancellationToken);
-            }
-
-            public Task<string> GetResourceIdentifierAsync(ITrace trace, CancellationToken cancellationToken)
-            {
-                return this.inner.GetResourceIdentifierAsync(trace, cancellationToken);
-            }
-
-            public Task MergeAsync(FeedRangeInternal feedRange1, FeedRangeInternal feedRange2, CancellationToken cancellationToken)
-            {
-                return this.inner.MergeAsync(feedRange1, feedRange2, cancellationToken);
-            }
-
-            public Task<TryCatch<ChangeFeedPage>> MonadicChangeFeedAsync(
-                FeedRangeState<ChangeFeedState> feedRangeState,
-                ChangeFeedPaginationOptions changeFeedPaginationOptions,
-                ITrace trace,
-                CancellationToken cancellationToken)
-            {
-                return this.inner.MonadicChangeFeedAsync(feedRangeState, changeFeedPaginationOptions, trace, cancellationToken);
-            }
-
-            public Task<TryCatch<Record>> MonadicCreateItemAsync(CosmosObject payload, CancellationToken cancellationToken)
-            {
-                return this.inner.MonadicCreateItemAsync(payload, cancellationToken);
-            }
-
-            public Task<TryCatch<List<FeedRangeEpk>>> MonadicGetChildRangeAsync(
-                FeedRangeInternal feedRange,
-                ITrace trace,
-                CancellationToken cancellationToken)
-            {
-                return this.inner.MonadicGetChildRangeAsync(feedRange, trace, cancellationToken);
-            }
-
-            public Task<TryCatch<List<FeedRangeEpk>>> MonadicGetFeedRangesAsync(ITrace trace, CancellationToken cancellationToken)
-            {
-                return this.inner.MonadicGetFeedRangesAsync(trace, cancellationToken);
-            }
-
-            public Task<TryCatch<string>> MonadicGetResourceIdentifierAsync(ITrace trace, CancellationToken cancellationToken)
-            {
-                return this.inner.MonadicGetResourceIdentifierAsync(trace, cancellationToken);
-            }
-
-            public Task<TryCatch> MonadicMergeAsync(
-                FeedRangeInternal feedRange1,
-                FeedRangeInternal feedRange2,
-                CancellationToken cancellationToken)
-            {
-                return this.inner.MonadicMergeAsync(feedRange1, feedRange2, cancellationToken);
-            }
-
-            public async Task<TryCatch<QueryPage>> MonadicQueryAsync(
-                SqlQuerySpec sqlQuerySpec,
-                FeedRangeState<QueryState> feedRangeState,
-                QueryPaginationOptions queryPaginationOptions,
-                ITrace trace,
-                CancellationToken cancellationToken)
-            {
-                TryCatch<QueryPage> queryPage = await this.inner.MonadicQueryAsync(sqlQuerySpec, feedRangeState, queryPaginationOptions, trace, cancellationToken);
-
-                if (queryPage.Failed)
-                {
-                    return queryPage;
-                }
-
-                QueryPage page = queryPage.Result;
-                return TryCatch<QueryPage>.FromResult(new QueryPage(
-                    page.Documents,
-                    page.RequestCharge,
-                    page.ActivityId,
-                    page.ResponseLengthInBytes,
-                    page.CosmosQueryExecutionInfo,
-                    page.DistributionPlanSpec,
-                    page.DisallowContinuationTokenMessage,
-                    page.AdditionalHeaders,
-                    page.State,
-                    streaming: false));
-            }
-
-            public Task<TryCatch<ReadFeedPage>> MonadicReadFeedAsync(
-                FeedRangeState<ReadFeedState> feedRangeState,
-                ReadFeedPaginationOptions readFeedPaginationOptions,
-                ITrace trace,
-                CancellationToken cancellationToken)
-            {
-                return this.inner.MonadicReadFeedAsync(feedRangeState, readFeedPaginationOptions, trace, cancellationToken);
-            }
-
-            public Task<TryCatch<Record>> MonadicReadItemAsync(
-                CosmosElement partitionKey,
-                string identifer,
-                CancellationToken cancellationToken)
-            {
-                return this.inner.MonadicReadItemAsync(partitionKey, identifer, cancellationToken);
-            }
-
-            public Task<TryCatch> MonadicRefreshProviderAsync(ITrace trace, CancellationToken cancellationToken)
-            {
-                return this.inner.MonadicRefreshProviderAsync(trace, cancellationToken);
-            }
-
-            public Task<TryCatch> MonadicSplitAsync(FeedRangeInternal feedRange, CancellationToken cancellationToken)
-            {
-                return this.inner.MonadicSplitAsync(feedRange, cancellationToken);
-            }
-
-            public async Task<QueryPage> QueryAsync(
-                SqlQuerySpec sqlQuerySpec,
-                FeedRangeState<QueryState> feedRangeState,
-                QueryPaginationOptions queryPaginationOptions,
-                ITrace trace,
-                CancellationToken cancellationToken)
-            {
-                TryCatch<QueryPage> queryPage = await this.MonadicQueryAsync(
-                    sqlQuerySpec,
-                    feedRangeState,
-                    queryPaginationOptions,
-                    trace,
-                    cancellationToken);
-                queryPage.ThrowIfFailed();
-                return queryPage.Result;
-            }
-
-            public Task<ReadFeedPage> ReadFeedAsync(
-                FeedRangeState<ReadFeedState> feedRangeState,
-                ReadFeedPaginationOptions readFeedPaginationOptions,
-                ITrace trace,
-                CancellationToken cancellationToken)
-            {
-                return this.inner.ReadFeedAsync(feedRangeState, readFeedPaginationOptions, trace, cancellationToken);
-            }
-
-            public Task<Record> ReadItemAsync(CosmosElement partitionKey, string identifier, CancellationToken cancellationToken)
-            {
-                return this.inner.ReadItemAsync(partitionKey, identifier, cancellationToken);
-            }
-
-            public Task RefreshProviderAsync(ITrace trace, CancellationToken cancellationToken)
-            {
-                return this.inner.RefreshProviderAsync(trace, cancellationToken);
-            }
-
-            public Task SplitAsync(FeedRangeInternal feedRange, CancellationToken cancellationToken)
-            {
-                return this.inner.SplitAsync(feedRange, cancellationToken);
-            }
         }
 
         private static async Task<IDocumentContainer> CreateDocumentContainerAsync(
